+++ CHANGELOG.md
@@ -18,10 +18,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added an example for using the new FSMC interface with the provided
   `display-interface` driver and the `st7789` driver on a F413Discovery board [#302]
 - Derive `Eq`, `PartialEq`, `Copy` and `Clone` for error types
+- Added open-drain pin mode support for PWM output [#313]
 
 [#265]: https://github.com/stm32-rs/stm32f4xx-hal/pull/265
 [#297]: https://github.com/stm32-rs/stm32f4xx-hal/pull/297
 [#302]: https://github.com/stm32-rs/stm32f4xx-hal/pull/302
+[#313]: https://github.com/stm32-rs/stm32f4xx-hal/pull/313
 
 ### Changed